Fix random problems with data that can't be sent to the server
Fact 1 : The browser limits the number of open sockets to the same server
Fact 2 : Nginx change the requests order when forwarded to TOMUSS server

TOMUSS does not display the feedback green square until previous requests
are received. So each orange square is an open socket.

The number of feedbacks square has been reduced from 10 to 5.