Add a link to display the display tree
The profiling link is now a single character so it is always
displayed even if debugging is not activated.

Beware, these debugging links are common to 'home' and 'suivi' pages:

D('Profiling'   , ['User', 'HomeTop'],3)
D('Tree'        , ['User', 'HomeTop'],4)

It is clearly not nice, the display-framework should be enhanced
to allow to define twice the same node?