Add an experimental invitation plugin
It allows to add new users login in the first column of a table.
The cell comment of the ID indicates the mail of the user
who has received the invitation.

An invitation can be used multiple times with different users.
The table master can see in the comment the user who shared
its invitation link.

The invitation link expired after the fixed number of days.
It is not possible to cancel an invitation without canceling
all invitations by changing the config_table random seed.