The public display of a table can now be with authentication
Beware: the cells are not modifiable client side, but server side
they are changeable if the user is allowed to.

To allow client side modification, the table must be modifiable (if True or)
and statistics must be sent to the client in order to know if
the value change is allowed. It is complex because 'repetition'
and 'columngroup' in an anonymous context may not be manageable
client side. The 'Enumeration' case is the complex one.