Change the average computation in the ambiguous rounding case
Rounding 7.75 to 0.5 will give 7.5

With the previous formula, the result was not the same in Python
and JavaScript and so in the table editor and 'suivi' page.

It is an important change, averages must not be changed
in the old tables (in production 0.06% of averages are changed).
So there are utilities to fix the value of old averages:

  * average_compatibility.py search for changes of computed values
  * backward_fix.py patch the database in order to avoid any change

Once a computed values fixed, the cell comment must be changed to 'free' it.